Compile with -ffast-math
-ffast-math trades away standards conformance and numerical precision
for speed. For example it truncates subnormal numbers to zero, and can
convert division into multiplication by the reciprical. I don't
believe these kinds of optimisations will hurt t-SNE, especially since
the implementation is full of ad-hoc optimisations anyway (not least
of which is the whole Barnes Hut thing).

It appears that this enables automatic SSE vectorisation in building
the tree with GCC 4.9 on x86-64 Linux.