Replace -msse2 with -march=native
On machines where SSE2 is supported, it will be enabled by
-march=native, but -march-native will also use more advanced
instruction sets if they are available.

On machines where SSE2 is not supported, compiling with -msse2 is not
going to work, while -march=native will.

The assumption is most users will compile on the target machine.

The last 4 commits have the following cumulative effect on the mnist
test:

    Building tree...
    Done in 1891.01 seconds (sparsity = 0.001854)!
    Learning embedding...
    Iteration 999: error is 3.509694
    Fitting performed in 689.22 seconds.

    Building tree...
    Done in 1565.03 seconds (sparsity = 0.001854)!
    Learning embedding...
    Iteration 999: error is 3.509400
    Fitting performed in 673.73 seconds.