Add option of cosine distance
It could be that C++ is a TERRIBLE language, or it could be that I just
don't know it well, but there seems to be no simple way of sharing the
tree building code between the cosine_distance and euclidean_distance
versions.

Relatively untested: I know it does *something*, as in lays the MNIST
dataset out in a peculiar spiral, but I don't know if that is the right
thing.