Cosine distance wasn't working, replace with normalisation, add test
I don't think the cosine distance worked very well with the vantage
point tree (because it is non-metric, no triangle inequality, etc).

Instead there is now an option to project all the vectors onto the
unit hypersphere, and use euclidean distance there. Intuitively,
things that are close in angle should be close on the hypersphere,
although exactly how well that works in high dimensions. It seems to
work reasonably well on a simple test.